changed timestamp to float

+++ gmap/models.py
@@ -36,7 +36,7 @@ class BandwidthManager(models.Manager):
 class Bandwidth(models.Model):
     """Bandwidth on an NDN network link"""
     link = models.IntegerField()
-    time = models.BigIntegerField()
+    time = models.FloatField()
     rx = models.BigIntegerField()
     tx = models.BigIntegerField()
     update_date = models.DateTimeField(default=datetime.now, editable=False)
